GreatFET One: schematic changes permitting power supplied by neighbor
Renamed VBUS on J2 pin 2 to +5V and isolated +5V from USB0_VBUS with D5,
a diode to protect the USB host.  Renamed U1 pin 21 from VBUS to
USB0_SENSE and added a votage divider (R25 and R26) to protect this pin
from voltages greater than 3.6 V in the unusual situation that
USB0_SENSE is present without VCC present.  (This happened and appears
to have resulted in damage when someone applied power to a home-built
unit that had an assembly error.)  Changed USB0_SENSE and USB1_SENSE
voltage divider resistor values so they work even when VBUS is as low as
4.4 V.  Added supply voltage requirements to notes in schematic.

These changes are not represented in the layout yet.